Cast pid_t to int to fix gcc-2.95.3 warning on solaris.

+++ src/main.c
@@ -307,7 +307,7 @@ int main(int argc, char *argv[])
 	if (pid_fd != -1) {
 		ftruncate(pid_fd, 0);
 		if (tuning.num_processes > 1)
-			sprintf(buf, "-%d\n", getpgrp());
+			sprintf(buf, "-%d\n", (int) getpgrp());
 		else
 			sprintf(buf, "%d\n", my_pid);
 		write(pid_fd, buf, strlen(buf));
